+++ p22/src/bin/main.rs
@@ -1,0 +1,5 @@
+use p22::song::print_twelve_days_of_christmas;
+
+fn main() {
+    print_twelve_days_of_christmas();
+}
